+++ internal/pkg/facade/certificates/delete_test.go
@@ -19,7 +19,7 @@ func TestFacade_DeleteCertificate(t *testing.T) {
 		fx.dbMock.ExpectBegin()
 		fx.dbMock.ExpectRollback()
 
-		fx.certificateStorage.EXPECT().GetCertificateByID(mock.Anything, int32(1)).Return(&database.Certificate{}, sql.ErrNoRows)
+		fx.certificateStorage.EXPECT().GetCertificateByID(mock.Anything, 1).Return(&database.Certificate{}, sql.ErrNoRows)
 
 		err := fx.facade.DeleteCertificate(fx.ctx, 1)
 
@@ -36,7 +36,7 @@ func TestFacade_DeleteCertificate(t *testing.T) {
 		fx.dbMock.ExpectBegin()
 		fx.dbMock.ExpectRollback()
 
-		fx.certificateStorage.EXPECT().GetCertificateByID(mock.Anything, int32(1)).Return(&database.Certificate{}, errors.New("some error"))
+		fx.certificateStorage.EXPECT().GetCertificateByID(mock.Anything, 1).Return(&database.Certificate{}, errors.New("some error"))
 
 		err := fx.facade.DeleteCertificate(fx.ctx, 1)
 
@@ -52,7 +52,7 @@ func TestFacade_DeleteCertificate(t *testing.T) {
 		fx.dbMock.ExpectBegin()
 		fx.dbMock.ExpectRollback()
 
-		fx.certificateStorage.EXPECT().GetCertificateByID(mock.Anything, int32(1)).Return(&database.Certificate{
+		fx.certificateStorage.EXPECT().GetCertificateByID(mock.Anything, 1).Return(&database.Certificate{
 			ID: 1,
 			DeletedAt: sql.NullTime{
 				Valid: true,
@@ -75,11 +75,11 @@ func TestFacade_DeleteCertificate(t *testing.T) {
 		fx.dbMock.ExpectBegin()
 		fx.dbMock.ExpectRollback()
 
-		fx.certificateStorage.EXPECT().GetCertificateByID(mock.Anything, int32(1)).Return(&database.Certificate{
+		fx.certificateStorage.EXPECT().GetCertificateByID(mock.Anything, 1).Return(&database.Certificate{
 			ID: 1,
 		}, nil)
 
-		fx.certificateStorage.EXPECT().DeleteCertificate(mock.Anything, int32(1)).Return(errors.New("some error"))
+		fx.certificateStorage.EXPECT().DeleteCertificate(mock.Anything, 1).Return(errors.New("some error"))
 
 		err := fx.facade.DeleteCertificate(fx.ctx, 1)
 
@@ -95,11 +95,11 @@ func TestFacade_DeleteCertificate(t *testing.T) {
 		fx.dbMock.ExpectBegin()
 		fx.dbMock.ExpectCommit()
 
-		fx.certificateStorage.EXPECT().GetCertificateByID(mock.Anything, int32(1)).Return(&database.Certificate{
+		fx.certificateStorage.EXPECT().GetCertificateByID(mock.Anything, 1).Return(&database.Certificate{
 			ID: 1,
 		}, nil)
 
-		fx.certificateStorage.EXPECT().DeleteCertificate(mock.Anything, int32(1)).Return(nil)
+		fx.certificateStorage.EXPECT().DeleteCertificate(mock.Anything, 1).Return(nil)
 
 		err := fx.facade.DeleteCertificate(fx.ctx, 1)
